Repair relations and delete parser manager class
